Updating types.

+++ src/agent/onefuzz-result/src/job_result.rs
@@ -34,7 +34,7 @@ struct JobResult {
     machine_id: Uuid,
     machine_name: String,
     data: Vec<JobResultData>,
-    value: HashMap<String, u64>,
+    value: HashMap<String, f64>,
 }
 
 #[derive(Clone)]
@@ -114,12 +114,12 @@ pub async fn init_job_result(
 
 #[async_trait]
 pub trait JobResultSender {
-    async fn send_direct(&self, data: JobResultData, value: HashMap<String, u64>) -> Result<()>;
+    async fn send_direct(&self, data: JobResultData, value: HashMap<String, f64>) -> Result<()>;
 }
 
 #[async_trait]
 impl JobResultSender for TaskJobResultClient {
-    async fn send_direct(&self, data: JobResultData, value: HashMap<String, u64>) -> Result<()> {
+    async fn send_direct(&self, data: JobResultData, value: HashMap<String, f64>) -> Result<()> {
         let task_id = self.context.state.task_id;
         let job_id = self.context.state.job_id;
         let machine_id = self.context.state.machine_id;
@@ -143,7 +143,7 @@ impl JobResultSender for TaskJobResultClient {
 
 #[async_trait]
 impl JobResultSender for Option<TaskJobResultClient> {
-    async fn send_direct(&self, data: JobResultData, value: HashMap<String, u64>) -> Result<()> {
+    async fn send_direct(&self, data: JobResultData, value: HashMap<String, f64>) -> Result<()> {
         match self {
             Some(client) => client.send_direct(data, value).await,
             None => Ok(()),
